+++ src/ft_memchr.c
@@ -6,13 +6,13 @@
 /*   By: gbetting <gbetting@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/11 00:36:36 by gbetting          #+#    #+#             */
-/*   Updated: 2024/04/11 00:36:43 by gbetting         ###   ########.fr       */
+/*   Updated: 2024/06/26 05:36:45 by gbetting         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memchr(const void *s, int c, size_t n)
+void	*ft_memchr(const void *s, int8_t c, size_t n)
 {
 	unsigned char	*str;
 	unsigned char	chr;
